fix arm64 compile

+++ havana/havok/hk_mopp/ivp_compact_mopp.hxx
@@ -65,7 +65,7 @@ public:
      *****************************************************************************/
     const hkMoppCode* get_compact_ledge_tree_root() const 
 	{
-		char* base = (char*)(((int)this) + this->offset_ledgetree_root);
+		char* base = (char*)(((intp)this) + this->offset_ledgetree_root);
 		return((const hkMoppCode*)base);
     }
