+++ havana/havok/hk_physics/constraint/ball_socket/ball_socket_constraint.cpp
@@ -22,7 +22,7 @@ class hk_Ball_Socket_Work
 			return addr;
 		}
 #else
-		static inline void *operator new (unsigned int size, void *addr){
+		static inline void *operator new (size_t size, void *addr){
 			return addr;
 		}
