+++ havana/havok/hk_physics/constraint/prismatic/prismatic_constraint.cpp
@@ -28,7 +28,7 @@ class hk_Prismatic_Work
 			return addr;
 		}
 #else
-		static inline void *operator new (unsigned int size, void *addr){
+		static inline void *operator new (size_t size, void *addr){
 			return addr;
 		}
