+++ havana/havok/hk_physics/constraint/pulley/pulley_constraint.cpp
@@ -19,7 +19,7 @@ class hk_Pulley_Work
 			return addr;
 		}
 #else
-		static inline void *operator new (unsigned int size, void *addr){
+		static inline void *operator new (size_t size, void *addr){
 			return addr;
 		}
