+++ ivp_surface_manager/ivp_compact_surface.hxx
@@ -52,7 +52,7 @@ public:
      *  Description:    INTERNAL METHOD
      *****************************************************************************/
     const IVP_Compact_Ledgetree_Node *get_compact_ledge_tree_root() const {
-	char *base = (char *)(((int)this) + this->offset_ledgetree_root);
+	char *base = (char *)(((intptr_t)this) + this->offset_ledgetree_root);
 	return((const IVP_Compact_Ledgetree_Node *)base);
     }
